pass the hostname to SASL for now.
The hostname is defaulted to the domain name of the JID.  The 100% correct
solution for GSSAPI w.r.t. SASL would be to pass the actual hostname determined
by a DNS-SRV lookup to Authen::SASL. But that would require a reimplementation
of the SRV logic implemented in tcp_connect of AnyEvent::Socket, which I'm not
going to do voluntarily. If someone needs this, and setting the 'host'
parameter of AnyEvent::XMPP::Connection explicitly is not enough, he is
encouraged to implement it and send a patch.